requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
python-dotenv==0.19.2
postgres==4.0